Move bicep unit tests to Areas
