+++ tests/Areas/BicepSchema/UnitTests/EntitySerializationTests.cs
@@ -5,7 +5,7 @@ using System.Text.Json;
 using AzureMcp.Areas.BicepSchema.Services.ResourceProperties.Entities;
 using Xunit;
 
-namespace AzureMcp.Tests.Services.Azure.BicepSchema;
+namespace AzureMcp.Tests.Areas.BicepSchema.UnitTests;
 
 public class EntitySerializationTests
 {
