+++ tests/Areas/BicepSchema/UnitTests/ParseResourceTests.cs
@@ -4,7 +4,7 @@
 using AzureMcp.Areas.BicepSchema.Services.Support;
 using Xunit;
 
-namespace AzureMcp.Tests.Services.Azure.BicepSchema;
+namespace AzureMcp.Tests.Areas.BicepSchema.UnitTests;
 
 public class ParseResourceTests
 {
